+++ CHANGELOG.md
@@ -9,8 +9,7 @@ All notable changes to this project will be documented in this file. It uses the
 ### Added
 * Now targeting .NET 7 as per the [.NET release lifecycle](https://dotnet.microsoft.com/en-us/platform/support/policy/dotnet-core)
 ### Changed
-* Updated mapping to 4.5.0.2
-* Refactoring for .NET 7 and moving download from extension to service
+* Updated mapping to 4.12.1.1
 ### Deprecated
 ### Removed
 ### Fixed